#72: Rewrote the Discord send queue to utilise LBQ's blocking
In the current active git version this thread was guaranteed to use a CPU
at 100% usage unless there was a message present to be sent, which is only
a minor amount of time. This commit pulls both the jda-check out of the
main loop, reorganizes some of the try-clauses and uses blocking of
LinkedBlockingQueue.

Signed-Off: Rens Rikkerink <email-redacted>