Add env variable lookup support for Discord token/server/channel
Lookup will be performed if the value in the config is empty.
The first fallback is the property (discordchat.token, discordchat.server, discordchat.channels) which is set with the -Ddiscordchat.server=... VM argument.
The second fallback (if no property value is provided) is the environment variable (DISCORDCHAT_TOKEN, etc.) which is specified though the command line prior to launching the server.

Closes #97